代码恢复

+++ src/main/java/com/wf/core/cache/CacheKey.java
@@ -5,26 +5,26 @@ package com.wf.core.cache;
  * @author Fe 2016年4月16日
  */
 public interface CacheKey {
-    int MINUTE_1 = 60 * 1;
-    int MINUTE_5 = MINUTE_1 * 5;
-    int MINUTE_10 = MINUTE_1 * 10;
-    int MINUTE_15 = MINUTE_1 * 15;
-    int MINUTE_30 = MINUTE_1 * 30;
-    int HOUR_1 = MINUTE_1 * 60;
-    int HOUR_2 = HOUR_1 * 2;
-    int HOUR_3 = HOUR_1 * 3;
-    int HOUR_5 = HOUR_1 * 5;
-    int HOUR_12 = HOUR_1 * 12;
-    int DAY_1 = HOUR_1 * 24;
-    int DAY_2 = DAY_1 * 2;
-    int DAY_3 = DAY_1 * 3;
-    int DAY_5 = DAY_1 * 5;
-    int DAY_15 = DAY_1 * 15;
-    int WEEK_1 = DAY_1 * 7;
-    int WEEK_2 = WEEK_1 * 2;
-    int MONTH_1 = DAY_1 * 30;
-    int MONTH_2 = MONTH_1 * 2;
-    int MONTH_3 = MONTH_1 * 3;
+    long MINUTE_1 = 60L * 1;
+    long MINUTE_5 = MINUTE_1 * 5;
+    long MINUTE_10 = MINUTE_1 * 10;
+    long MINUTE_15 = MINUTE_1 * 15;
+    long MINUTE_30 = MINUTE_1 * 30;
+    long HOUR_1 = MINUTE_1 * 60;
+    long HOUR_2 = HOUR_1 * 2;
+    long HOUR_3 = HOUR_1 * 3;
+    long HOUR_5 = HOUR_1 * 5;
+    long HOUR_12 = HOUR_1 * 12;
+    long DAY_1 = HOUR_1 * 24;
+    long DAY_2 = DAY_1 * 2;
+    long DAY_3 = DAY_1 * 3;
+    long DAY_5 = DAY_1 * 5;
+    long DAY_15 = DAY_1 * 15;
+    long WEEK_1 = DAY_1 * 7;
+    long WEEK_2 = WEEK_1 * 2;
+    long MONTH_1 = DAY_1 * 30;
+    long MONTH_2 = MONTH_1 * 2;
+    long MONTH_3 = MONTH_1 * 3;
 
     public String key();
 }
